+++ src/components/ARPlayer/ARPlayer.tsx
@@ -5,7 +5,6 @@ import useAR from "./useAR";
 const ARPlayer = () => {
   const videoRef = useRef<HTMLVideoElement>(null);
   const { ARStream } = useAR();
-
   const { setStream } = useContext(SocketContext);
 
   useEffect(() => {
@@ -25,7 +24,7 @@ const ARPlayer = () => {
       autoPlay
       playsInline
       poster="https://www.pngfind.com/pngs/m/610-6104451_image-placeholder-png-user-profile-placeholder-image-png.png"
-      // muted
+      muted
       style={{
         width: "100%",
         objectFit: "cover",
