--- src/components/notification/Notification.tsx
@@ -1,22 +1,0 @@
-import { Button } from "@material-ui/core";
-import React, { useContext } from "react";
-import { SocketContext } from "src/pages/Context/Context";
-
-const Notification = () => {
-  const { answerCall, call, callAccepted } = useContext(SocketContext);
-
-  return (
-    <>
-      {call?.isReceivingCall && !callAccepted && (
-        <div style={{ display: "flex", justifyContent: "space-around" }}>
-          <h1>{call?.name} is calling:</h1>
-          <Button variant="contained" color="primary" onClick={answerCall}>
-            Answer
-          </Button>
-        </div>
-      )}
-    </>
-  );
-};
-
-export default Notification;
