+++ src/pages/Conference/Conference.tsx
@@ -2,10 +2,9 @@
 import { Grid } from "@material-ui/core";
 import { useContext, useEffect } from "react";
 import ARPlayer from "src/components/ARPlayer/ARPlayer";
+import useSocket from "src/components/ARPlayer/useSocket";
 import BottomBar from "src/components/BottomComponents/Bottombar";
 import VideoPlayer from "src/components/videoplayer/VideoPlayer";
-import { useTray } from "src/hooks/useTray";
-import { useWebcam } from "src/hooks/useWebcam";
 import { SocketContext } from "src/pages/Context/Context";
 
 // const useStyles = makeStyles((theme: Theme) =>
@@ -24,30 +23,14 @@ import { SocketContext } from "src/pages/Context/Context";
 const Conference = () => {
   // const classes = useStyles();
 
-  const context = useContext(SocketContext);
-  const { stream, otherStreams } = context!;
-
-  // const { id } = useParams<{ id: string }>();
-
-  const { setState, usingAR } = useTray();
-  const { videoTracks } = useWebcam();
-
   useEffect(() => {
-    setState({ type: "webcam", value: true });
-    setState({ type: "microphone", value: true });
-
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
+  useSocket();
+  const context = useContext(SocketContext);
+  const { otherStreams } = context!;
 
-  useEffect(() => {
-    const videoObj = videoTracks ? new MediaStream(videoTracks) : null;
-
-    // console.log("What");
-    if (!usingAR) context.setStream(videoObj!);
-
-    // console.log(usingAR);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [usingAR, videoTracks]);
+  // const { id } = useParams<{ id: string }>();
 
   return (
     <Grid
@@ -66,20 +49,11 @@ const Conference = () => {
         alignItems="center"
         justify="center"
       >
-        {/* <ThreexComp /> */}
-        {usingAR ? (
-          <Grid item xs={8} md={6} lg={4}>
-            <ARPlayer />
-          </Grid>
-        ) : (
-          <Grid item xs={8} md={6} lg={4}>
-            <VideoPlayer stream={stream!} muted={true} />
-          </Grid>
-        )}
+        <Grid item xs={8} md={6} lg={4}>
+          <ARPlayer />
+        </Grid>
 
         {otherStreams?.map((otherStream) => {
-          console.log("Other Stream");
-          console.log(otherStream);
           return (
             <Grid item xs={12} md={6} lg={4}>
               <VideoPlayer
